+++ src/tasks/Task_24.java
@@ -3,9 +3,9 @@ package tasks;
 public class Task_24 {
 
     public static void main(String[] args) {
-        double a = 2;
-        double b = 2;
-        double c = 6;
+        int a = 2;
+        int b = 2;
+        int c = 6;
         if (a == 2 && b == 2 || b == 2 && c == 2 || a == 2 && c == 2) {
             System.out.println(true);
         } else {
@@ -14,9 +14,9 @@ public class Task_24 {
     }
 
     /*public static void main(String[] args) {
-        double a = 2;
-        double b = 27;
-        double c = 6;
+        int a = 2;
+        int b = 27;
+        int c = 6;
 
         if (a == 2) {
             if (b == 2) {
